+++ backend/src/config/datasource.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './entity/User';
+import { User } from '../entity/User';
 
 @Module({
   imports: [
